+++ Sstv.DomainExceptions/SourceGenerators/Constants.cs
@@ -1,0 +1,35 @@
+namespace Sstv.DomainExceptions.SourceGenerators;
+
+/// <summary>
+/// Constants.
+/// </summary>
+internal sealed class Constants
+{
+    /// <summary>
+    /// Library main namespace.
+    /// </summary>
+    public const string NAMESPACE = "Sstv.DomainExceptions";
+
+    /// <summary>
+    /// Attribute class name, that should be placed on top of enum with error codes.
+    /// </summary>
+    public const string ATTRIBUTE_CLASS_NAME = "ErrorDescription";
+
+    /// <summary>
+    /// Full name of <see cref="ATTRIBUTE_CLASS_NAME"/>.
+    /// </summary>
+    public const string ATTRIBUTE_FULL_NAME = NAMESPACE + "." + ATTRIBUTE_CLASS_NAME + "Attribute";
+
+    /// <summary>
+    /// Error description holder class name.
+    /// </summary>
+    public const string ERROR_DESCRIPTION_CLASS_NAME = "ErrorDescription";
+
+    public static class NamedArguments
+    {
+        public const string HELP_LINK = "HelpLink";
+        public const string PREFIX = "Prefix";
+        public const string DESCRIPTION = "Description";
+        public const string ERROR_CODE_LENGTH = "ErrorCodeLength";
+    }
+}