+++ Sstv.DomainExceptions.Extensions.SerilogEnricher/CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-02-11
+
+- Source gen release
+
+
 ## [1.0.0] - 2023-10-04
 
 ### Added